backend/requirements.txt
pydantic==2.6.1
pydantic_core==2.16.2
pyhumps==3.8.0
python-dotenv==1.0.1
Quart==0.19.4
quart-schema==0.19.0